chore: move processorType to windows tools
