+++ packages/cli/src/tools/windows/__tests__/androidWinHelpers.test.ts
@@ -1,6 +1,6 @@
 import * as androidWinHelpers from '../androidWinHelpers';
 import * as executeWinCommand from '../executeWinCommand';
-import * as processorType from '../../processorType';
+import * as processorType from '../processorType';
 import {ExecaChildProcess} from 'execa';
 
 describe('androidWinHelpers', () => {
